refactor(frontend): change admin directories to teacher directories
