+++ services/frontend/src/components/courses/teacher/course-cell-status.tsx
