+++ services/frontend/src/components/post/teacher/editor.tsx
