+++ services/frontend/src/components/post/teacher/preview.tsx
